vm_core module.
