+++ crates/motoko/src/lib/mod.rs
@@ -27,6 +27,7 @@ mod serde_utils;
 pub mod shared;
 pub mod value;
 pub mod vm;
+pub mod vm_core;
 pub mod vm_ops;
 pub mod vm_types;
